+++ src/logger.rs
@@ -1,4 +1,5 @@
 use std::env;
+use std::sync::atomic::{AtomicBool, Ordering};
 
 use once_cell::sync::Lazy;
 
@@ -27,4 +28,12 @@ pub fn warn(message: &str) {
 
 pub fn info(message: &str) {
     eprintln!("INFO: {message}");
+}
+
+pub fn warn_deprecated_notifications() {
+    static HAS_WARNED: AtomicBool = AtomicBool::new(false);
+    let before = HAS_WARNED.swap(true, Ordering::Relaxed);
+    if !before {
+        warn("Notifications are deprecated and will be replaced by a simpler machine-readable output. This way you will be able to control the exact notifications even better yourself. The details on this are not yet finalized. Please join the discussion on https://github.com/EdJoPaTo/website-stalker/discussions/172");
+    }
 }