+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - RSS: remove website-stalker version from the generator field
 - Instantly panic or print cleaner human error message
+- Deprecate notifications
 
 ## [0.22.0] - 2024-02-13